Added new div
The way I did the divs, there was a grey background for the h2 but
nothing below it. I thought I fixed it, but the way I did the divs made
it so there was a white space between the h2 and the rest of the
content below it. I added a div that went around everything and added
the section-mid tag to it.